refactor(react-utils): import mergeRefs from react-utils
