+++ packages/breadcrumb/package.json
@@ -56,6 +56,7 @@
     "react": "^17.0.1"
   },
   "dependencies": {
+    "@chakra-ui/react-utils": "1.0.0",
     "@chakra-ui/utils": "1.4.0"
   },
   "peerDependencies": {
