+++ packages/breadcrumb/src/breadcrumb.tsx
@@ -10,7 +10,8 @@ import {
   useStyles,
   HTMLChakraProps,
 } from "@chakra-ui/system"
-import { cx, getValidChildren, __DEV__ } from "@chakra-ui/utils"
+import { cx, __DEV__ } from "@chakra-ui/utils"
+import { getValidChildren } from "@chakra-ui/react-utils"
 import * as React from "react"
 
 export interface BreadcrumbSeparatorProps extends HTMLChakraProps<"div"> {
