+++ packages/hooks/package.json
@@ -50,6 +50,7 @@
     "lint:types": "tsc --noEmit"
   },
   "dependencies": {
+    "@chakra-ui/react-utils": "1.0.0",
     "@chakra-ui/utils": "1.4.0",
     "compute-scroll-into-view": "1.0.14",
     "copy-to-clipboard": "3.3.1"
