+++ packages/input/package.json
@@ -54,6 +54,7 @@
   },
   "dependencies": {
     "@chakra-ui/form-control": "1.2.3",
+    "@chakra-ui/react-utils": "1.0.0",
     "@chakra-ui/utils": "1.4.0"
   },
   "devDependencies": {
