+++ packages/layout/package.json
@@ -59,6 +59,7 @@
   },
   "dependencies": {
     "@chakra-ui/icon": "1.1.3",
+    "@chakra-ui/react-utils": "1.0.0",
     "@chakra-ui/utils": "1.4.0"
   },
   "devDependencies": {
