+++ packages/modal/src/use-modal.ts
@@ -1,5 +1,6 @@
 import { useIds } from "@chakra-ui/hooks"
-import { callAllHandlers, mergeRefs, PropGetter } from "@chakra-ui/utils"
+import { callAllHandlers, PropGetter } from "@chakra-ui/utils"
+import { mergeRefs } from "@chakra-ui/react-utils"
 import { hideOthers, Undo } from "aria-hidden"
 import {
   KeyboardEvent,
