+++ packages/number-input/src/use-number-input.ts
@@ -13,13 +13,13 @@ import {
   isBrowser,
   isNull,
   maxSafeInteger,
-  mergeRefs,
   minSafeInteger,
   normalizeEventKey,
   PropGetter,
   StringOrNumber,
   withFlushSync,
 } from "@chakra-ui/utils"
+import { mergeRefs } from "@chakra-ui/react-utils"
 import * as React from "react"
 import { useSpinner } from "./use-spinner"
 import {
