+++ packages/popover/src/use-popover.ts
@@ -21,10 +21,10 @@ import {
   getRelatedTarget,
   HTMLProps,
   isBrowser,
-  mergeRefs,
   PropGetter,
   px,
 } from "@chakra-ui/utils"
+import { mergeRefs } from "@chakra-ui/react-utils"
 import { RefObject, useCallback, useEffect, useRef } from "react"
 
 const TRIGGER = {
