+++ packages/radio/src/use-radio.ts
@@ -3,11 +3,11 @@ import {
   ariaAttr,
   callAllHandlers,
   dataAttr,
-  mergeRefs,
   pick,
   PropGetter,
   warn,
 } from "@chakra-ui/utils"
+import { mergeRefs } from "@chakra-ui/react-utils"
 import { visuallyHiddenStyle } from "@chakra-ui/visually-hidden"
 import {
   ChangeEvent,
